+++ lib/view_component/engine.rb
@@ -27,7 +27,6 @@ module ViewComponent
         options[config_option] ||= ViewComponent::Base.public_send(config_option)
       end
       options.instrumentation_enabled = false if options.instrumentation_enabled.nil?
-      options.render_monkey_patch_enabled = true if options.render_monkey_patch_enabled.nil?
       options.show_previews = (Rails.env.development? || Rails.env.test?) if options.show_previews.nil?
 
       if options.show_previews
@@ -83,46 +82,6 @@ module ViewComponent
       ActiveSupport.on_load(:after_initialize) do
         ViewComponent::Base.descendants.each(&:compile) if Rails.application.config.eager_load
       end
-    end
-
-    initializer "view_component.monkey_patch_render" do |app|
-      next if Rails.version.to_f >= 6.1 || !app.config.view_component.render_monkey_patch_enabled
-
-      # :nocov:
-      ViewComponent::Deprecation.deprecation_warning("Monkey patching `render`", "ViewComponent 4.0 will remove the `render` monkey patch")
-
-      ActiveSupport.on_load(:action_view) do
-        require "view_component/render_monkey_patch"
-        ActionView::Base.prepend ViewComponent::RenderMonkeyPatch
-      end
-
-      ActiveSupport.on_load(:action_controller) do
-        require "view_component/rendering_monkey_patch"
-        require "view_component/render_to_string_monkey_patch"
-        ActionController::Base.prepend ViewComponent::RenderingMonkeyPatch
-        ActionController::Base.prepend ViewComponent::RenderToStringMonkeyPatch
-      end
-      # :nocov:
-    end
-
-    initializer "view_component.include_render_component" do |_app|
-      next if Rails.version.to_f >= 6.1
-
-      # :nocov:
-      ViewComponent::Deprecation.deprecation_warning("using `render_component`", "ViewComponent 4.0 will remove `render_component`")
-
-      ActiveSupport.on_load(:action_view) do
-        require "view_component/render_component_helper"
-        ActionView::Base.include ViewComponent::RenderComponentHelper
-      end
-
-      ActiveSupport.on_load(:action_controller) do
-        require "view_component/rendering_component_helper"
-        require "view_component/render_component_to_string_helper"
-        ActionController::Base.include ViewComponent::RenderingComponentHelper
-        ActionController::Base.include ViewComponent::RenderComponentToStringHelper
-      end
-      # :nocov:
     end
 
     initializer "static assets" do |app|