+++ test/sandbox/test/config_test.rb
@@ -15,7 +15,6 @@ module ViewComponent
       assert_equal @config.show_previews_source, false
       assert_equal @config.instrumentation_enabled, false
       assert_equal @config.use_deprecated_instrumentation_name, true
-      assert_equal @config.render_monkey_patch_enabled, true
       assert_equal @config.show_previews, true
       assert_equal @config.preview_paths, ["#{Rails.root}/test/components/previews"]
     end